Rename L2Action to ActionWithL2.js
