+++ src/actions/aaveV3/AaveV3CollateralSwitchAction.js
@@ -1,10 +1,10 @@
-const L2Action = require("../../L2Action");
+const ActionWithL2 = require("../../ActionWithL2");
 const { getAddr } = require('../../addresses.js');
 
 /**
  *  AaveV3CollateralSwitchAction - Aave enable/disable token usage as collateral for AaveV3 position
  */
-class AaveV3CollateralSwitchAction extends L2Action {
+class AaveV3CollateralSwitchAction extends ActionWithL2 {
   /**
    * @param useDefaultMarket {boolean} If this is true it defaults to the hardcoded market in contract
    * @param market {EthAddress} Address provider for specific market
